requirements.txt
# Vision-IoS Python Dependencies

# Database
psycopg2-binary==2.9.9

# Environment variables
python-dotenv==1.0.0

# Encryption for agent keys
cryptography==43.0.3

# LLM (Anthropic Claude)
anthropic>=0.74.0

# Binance API
python-binance==1.0.19

# Optional: Data analysis
pandas==2.1.4
numpy==1.26.2
